add helper function prepareGeneratedFile
